fix: split config
- rename AppConfig.java -> DBConfig.java
- create OpmlConfig.java
